Confirmed bots are learning and spreading good traits in simple navigation test.
Did quite a bit of debugging and have verified the networks is getting inputs, giving outputs, and trying to maximize fitness. In simple navigation (maximizing forward movement) learning is happening, but not perfect. Move X and Move Y can only be positive and Yaw can only be in one direction. These need to be fixed, which should be straightforward